Add 'modcontent' to 'mods'
Drops the release-20171014 'modcontent' folder into the 'mod' folder.

Enables user to download neccessary assets with the repository. Without the folder the game crashes when selecting 'Manage Content' in the Main Menu.

Not sure if I missed a step with the initial repository or it  being automatically added with packaging and user installation.

As of now the Mod Content Manager is added as-is from the release-20171014 source code.